+++ attributes/default.rb
@@ -87,3 +87,5 @@ default['slurm']['custom_template_banner'] = "
 "
 
 default['slurm']['user'] = default['slurm']['group'] = 'slurm'
+default['slurm']['uid'] = default['slurm']['gid'] = 999
+default['slurm']['comment'] = 'Slurm Workload Manager'